requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
google-generativeai==0.8.0
twilio==9.3.0
bcrypt==4.2.0
aiofiles==24.1.0
python-dotenv==1.0.1
itsdangerous==2.2.0
